# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "transac",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/stores": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "List all stores",
        "operationId": "list_stores",
        "responses": {
          "200": {
            "description": "List of stores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoresListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Create a new store",
        "operationId": "create_store",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Store created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/stores/{id}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "Get a store by ID",
        "operationId": "get_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "404": {
            "description": "Store not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "summary": "Update a store",
        "operationId": "update_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Store updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input"
          },
          "404": {
            "description": "Store not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "summary": "Delete a store",
        "operationId": "delete_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Store deleted successfully"
          },
          "404": {
            "description": "Store not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/stores/{id}/share": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "Generate store sharing links",
        "operationId": "get_store_share_links",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store sharing links generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreShareResponse"
                }
              }
            }
          },
          "404": {
            "description": "Store not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateStoreRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "contact_phone": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Store": {
        "$ref": "#/components/schemas/crate.entity.store.Model"
      },
      "StoreResponse": {
        "type": "object",
        "required": [
          "store"
        ],
        "properties": {
          "store": {
            "$ref": "#/components/schemas/Store"
          }
        }
      },
      "StoreShareResponse": {
        "type": "object",
        "required": [
          "store_id",
          "share_url",
          "whatsapp_share_url"
        ],
        "properties": {
          "share_url": {
            "type": "string"
          },
          "store_id": {
            "type": "string"
          },
          "whatsapp_share_url": {
            "type": "string"
          }
        }
      },
      "StoresListResponse": {
        "type": "object",
        "required": [
          "stores"
        ],
        "properties": {
          "stores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Store"
            }
          }
        }
      },
...
      "UpdateStoreRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "contact_phone": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}